Move paste functions out of insert mode module
